cleanup: move fs helpers into escapingfs
